Update README.md
a social platform designed for travelers, immigrants, international students, and expats. Here, you can meet, chat, network, date, and connect with people from all over the world. Whether you are in a new or foreign country, or simply looking to broaden your horizons, Visas Meet is the place for you.